feat(rpc): use uint32 instead of int64 grpc-ecosystem/grpc-gateway#438

+++ campusapis/teaching/v1/schedule.go
@@ -30,8 +30,8 @@ func (x *ScheduleItem) SetTime(q *CourseQuery) {
 	if err != nil {
 		return
 	}
-	x.StartTime = q.SectionReader.StartTime(d, FirstOfArray(x.Section)).Unix()
-	x.EndTime = q.SectionReader.EndTime(d, LastOfArray(x.Section)).Unix()
+	x.StartTime = uint32(q.SectionReader.StartTime(d, FirstOfArray(x.Section)).Unix())
+	x.EndTime = uint32(q.SectionReader.EndTime(d, LastOfArray(x.Section)).Unix())
 	x.IsThisWeek = true
 }
